+++ src/database/section.rs
@@ -1,5 +1,7 @@
 use super::{Line, Reference};
+use crate::{Issue, Location};
 use heck::ToKebabCase;
+use std::path::Path;
 
 /// a section in a document, from one heading to above the next heading
 #[derive(Debug, Eq, Hash, PartialEq)]
@@ -25,6 +27,35 @@ impl Section {
     /// provides the link anchor for this section, in GitHub format
     pub fn anchor(&self) -> String {
         format!("#{}", self.human_title().to_kebab_case())
+    }
+
+    /// populates the given issues list if this section has no content
+    pub fn check_empty(&self, path: &Path, issues: &mut Vec<Issue>) {
+        if !self.body.iter().any(|line| !line.text.is_empty()) {
+            issues.push(Issue::EmptySection {
+                location: Location {
+                    file: path.into(),
+                    line: self.line_number,
+                    start: 0,
+                    end: self.title_line.text.len() as u32,
+                },
+                title: self.human_title().into(),
+            });
+        }
+    }
+
+    /// populates the given issues list if this section has an empty title
+    pub fn check_empty_title(&self, path: &Path, issues: &mut Vec<Issue>) {
+        if self.human_title().is_empty() {
+            issues.push(Issue::SectionWithoutHeader {
+                location: Location {
+                    file: path.into(),
+                    line: self.line_number,
+                    start: 0,
+                    end: self.title_text_end(),
+                },
+            });
+        }
     }
 
     /// provides the cursor column at which the title text ends,
@@ -191,6 +222,123 @@ mod tests {
         for (give, want) in tests {
             let section = Section::with_title(give);
             assert_eq!(section.anchor(), want);
+        }
+    }
+    mod check_empty {
+        use crate::database::Document;
+        use crate::{Issue, Location};
+        use indoc::indoc;
+        use std::path::PathBuf;
+
+        #[test]
+        fn empty_section() {
+            let content = indoc! {"
+            # test document
+
+            ### empty section
+            ### next section
+
+            content"};
+            let doc = Document::from_str("test.md", content).unwrap();
+            let mut have = vec![];
+            for section in doc.content_sections {
+                section.check_empty(&PathBuf::from("test.md"), &mut have);
+            }
+            let want = vec![Issue::EmptySection {
+                location: Location {
+                    file: PathBuf::from("test.md"),
+                    line: 2,
+                    start: 0,
+                    end: 17,
+                },
+                title: "empty section".into(),
+            }];
+            pretty::assert_eq!(have, want);
+        }
+
+        #[test]
+        fn blank_line() {
+            let content = indoc! {"
+            # test document
+
+            ### empty section
+
+            ### next section
+
+            content"};
+            let doc = Document::from_str("test.md", content).unwrap();
+            let mut have = vec![];
+            for section in doc.content_sections {
+                section.check_empty(&PathBuf::from("test.md"), &mut have);
+            }
+            let want = vec![Issue::EmptySection {
+                location: Location {
+                    file: PathBuf::from("test.md"),
+                    line: 2,
+                    start: 0,
+                    end: 17,
+                },
+                title: "empty section".into(),
+            }];
+            pretty::assert_eq!(have, want);
+        }
+
+        #[test]
+        fn content() {
+            let content = indoc! {"
+            # test document
+
+            ### section with content
+
+            content"};
+            let doc = Document::from_str("test.md", content).unwrap();
+            let mut have = vec![];
+            for section in doc.content_sections {
+                section.check_empty(&PathBuf::from("test.md"), &mut have);
+            }
+            assert!(have.is_empty());
+        }
+    }
+
+    mod check_empty_title {
+        use crate::database::Document;
+        use crate::{Issue, Location};
+        use indoc::indoc;
+        use std::path::PathBuf;
+
+        #[test]
+        fn empty_title() {
+            let content = indoc! {"
+            # test document
+
+            ###
+            content
+            ###
+            content"};
+            let doc = Document::from_str("test.md", content).unwrap();
+            let mut have = vec![];
+            for section in doc.sections() {
+                section.check_empty_title(&PathBuf::from("test.md"), &mut have);
+            }
+            let want = vec![
+                Issue::SectionWithoutHeader {
+                    location: Location {
+                        file: PathBuf::from("test.md"),
+                        line: 2,
+                        start: 0,
+                        end: 3,
+                    },
+                },
+                Issue::SectionWithoutHeader {
+                    location: Location {
+                        file: PathBuf::from("test.md"),
+                        line: 4,
+                        start: 0,
+                        end: 3,
+                    },
+                },
+            ];
+            pretty::assert_eq!(have, want);
         }
     }
